Fixed ref test
